+++ packages/solid/src/select/index.ts
@@ -2,10 +2,10 @@ import { Select as SelectRoot, type SelectProps } from './select'
 import { SelectClearTrigger, type SelectClearTriggerProps } from './select-clear-trigger'
 import { SelectContent, type SelectContentProps } from './select-content'
 import { useSelectContext } from './select-context'
-import { SelectIndicator, type SelectIndicatorProps } from './select-indicator'
 import { SelectItem, type SelectItemProps } from './select-item'
 import { SelectItemGroup, type SelectItemGroupProps } from './select-item-group'
 import { SelectItemGroupLabel, type SelectItemGroupLabelProps } from './select-item-group-label'
+import { SelectItemIndicator, type SelectItemIndicatorProps } from './select-item-indicator'
 import { SelectLabel, type SelectLabelProps } from './select-label'
 import { SelectPositioner, type SelectPositionerProps } from './select-positioner'
 import { SelectTrigger, type SelectTriggerProps } from './select-trigger'
@@ -15,10 +15,10 @@ const Select = Object.assign(SelectRoot, {
   Root: SelectRoot,
   ClearTrigger: SelectClearTrigger,
   Content: SelectContent,
-  Indicator: SelectIndicator,
   Item: SelectItem,
   ItemGroup: SelectItemGroup,
   ItemGroupLabel: SelectItemGroupLabel,
+  ItemIndicator: SelectItemIndicator,
   Label: SelectLabel,
   Positioner: SelectPositioner,
   Trigger: SelectTrigger,
@@ -28,10 +28,10 @@ export {
   Select,
   SelectClearTrigger,
   SelectContent,
-  SelectIndicator,
   SelectItem,
   SelectItemGroup,
   SelectItemGroupLabel,
+  SelectItemIndicator,
   SelectLabel,
   SelectPositioner,
   SelectTrigger,
@@ -42,9 +42,9 @@ export {
 export type {
   SelectClearTriggerProps,
   SelectContentProps,
-  SelectIndicatorProps,
   SelectItemGroupLabelProps,
   SelectItemGroupProps,
+  SelectItemIndicatorProps,
   SelectItemProps,
   SelectLabelProps,
   SelectPositionerProps,