+++ packages/vue/src/select/select-clear-trigger.tsx
@@ -2,15 +2,15 @@ import { defineComponent } from 'vue'
 import { ark, type HTMLArkProps } from '../factory'
 import { useSelectContext } from './select-context'
 
-export type SelectTriggerProps = HTMLArkProps<'button'>
+export type SelectClearTriggerProps = HTMLArkProps<'button'>
 
-export const SelectTrigger = defineComponent({
-  name: 'SelectTrigger',
+export const SelectClearTrigger = defineComponent({
+  name: 'SelectClearTrigger',
   setup(_, { slots, attrs }) {
     const api = useSelectContext()
 
     return () => (
-      <ark.button {...api.value.triggerProps} {...attrs}>
+      <ark.button {...api.value.clearTriggerProps} {...attrs}>
         {slots.default?.()}
       </ark.button>
     )