+++ packages/vue/src/select/use-select.ts
@@ -1,43 +1,39 @@
-import { connect, machine } from '@zag-js/select'
-import { normalizeProps, useMachine } from '@zag-js/vue'
-import { computed, reactive, type ExtractPropTypes } from 'vue'
+import * as select from '@zag-js/select'
+import { normalizeProps, useMachine, type PropTypes } from '@zag-js/vue'
+import { computed, type ComputedRef } from 'vue'
 import { useEnvironmentContext } from '../environment'
-import { useId } from '../utils'
-import type { SelectContext } from './select'
+import type { Optional } from '../types'
+import { generateEventMap, useId } from '../utils'
+import { emits } from './select.props'
 
-export const useSelect = <T extends ExtractPropTypes<SelectContext>>(
-  emit: CallableFunction,
-  context: T,
-) => {
-  const reactiveContext = reactive(context)
+export type UseSelectProps = Optional<select.Context, 'id'> & {
+  modelValue?: select.Context['value']
+}
+export type UseSelectReturn = ComputedRef<select.Api<PropTypes>>
 
+export const useSelect = (props: UseSelectProps, emit: CallableFunction): UseSelectReturn => {
   const getRootNode = useEnvironmentContext()
+  const context = computed(() => {
+    const { modelValue, ...rest } = props
+    return {
+      ...rest,
+      value: modelValue,
+    }
+  })
+  const eventMap = generateEventMap(emits, emit)
 
   const [state, send] = useMachine(
-    machine({
-      ...reactiveContext,
-      id: reactiveContext.id || useId().value,
+    select.machine({
+      ...context.value,
+      id: context.value.id ?? useId().value,
       getRootNode,
-      selectedOption: reactiveContext.modelValue ?? reactiveContext.selectedOption,
+      ...eventMap,
       onChange: (details) => {
-        emit('change', { ...details })
-        emit('update:modelValue', { ...details })
-      },
-      onHighlight(details) {
-        emit('highlight', { ...details })
-      },
-      onClose() {
-        emit('close')
-      },
-      onOpen() {
-        emit('open')
+        emit('change', details.value)
+        emit('update:modelValue', details.value)
       },
     }),
+    { context },
   )
-
-  const api = computed(() => connect(state.value, send, normalizeProps))
-
-  return api
+  return computed(() => select.connect(state.value, send, normalizeProps))
 }
-
-export type UseSelectReturn = ReturnType<typeof useSelect>